requirements.txt
Flask
icalendar
PyGithub
pytz
gunicorn